+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 ### Fixed
 - Language variable boolean value is now correctly parsed in VariablesImport script
+- Ensure document object content always contains at laest one paragraph
 
 ### Removed